corrected unit tests
