+++ src/main/java/ma/nttdata/externals/module/interview/dto/GenerateInterviewQuestionsRequest.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.UUID;
 
 public record GenerateInterviewQuestionsRequest(
-        String promptCode,
         List<UUID> evaluationTypesIds
 ) {
 }
