--- src/test/java/ma/nttdata/externals/module/interview/service/impl/InterviewEvaluationUtilServImplTest.java
@@ -1,241 +1,0 @@
-package ma.nttdata.externals.module.interview.service.impl;
-
-import ma.nttdata.externals.module.candidate.dto.CandidateDTO;
-import ma.nttdata.externals.module.interview.dto.*;
-import ma.nttdata.externals.module.interview.entity.Evaluation;
-import ma.nttdata.externals.module.interview.entity.EvaluationType;
-import ma.nttdata.externals.module.interview.entity.Interview;
-import ma.nttdata.externals.module.interview.mapper.EvaluationMapper;
-import ma.nttdata.externals.module.interview.service.EvaluationServ;
-import ma.nttdata.externals.module.interview.service.InterviewServ;
-import ma.nttdata.externals.module.offer.dto.OfferDTO;
-import ma.nttdata.externals.module.prompt.dto.PromptDTO;
-import ma.nttdata.externals.module.prompt.service.PromptService;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.UUID;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-@ExtendWith(MockitoExtension.class)
-public class InterviewEvaluationUtilServImplTest {
-
-    @Mock
-    private EvaluationServ evaluationServ;
-    @Mock
-    private InterviewServ interviewServ;
-    @Mock
-    private EvaluationMapper evaluationMapper;
-    @Mock
-    private PromptService promptServ;
-
-    private InterviewEvaluationUtilServImpl interviewEvaluationUtilServ;
-
-    private static String JSON_SCHEMA = """
-            [
-              {
-                "score": "Double - between 0.00 and 100.00",
-                "feedback": "String - Try to give an overall feedback of the performance of the candidate in this evaluation type",
-                "evaluationTypeDescription": "String - use the exact 'description' field value from the corresponding EvaluationType entity"
-              }
-            ]
-            """;
-    private static final String INTERVIEW_EVALUATION_PROMPT = """
-            You're an expert interviewing manager and talent acquisition specialist.
-            We've passed an interview for an #offer, to a #candidate, and we've gathered the information output and prepared a list of #Question/#answer from that interview,
-            I will provide you below the needed information for them.
-            Prepare a list of evaluations for that interview, each #evaluation_type is an entry in this list, I will also give you the list of #evaluation_types that we need to evaluate this candidate in.
-                                       
-            #Take in consideration these instructions:
-             - The evaluations must be comprehensive and fair considering the job requirements in #offer_data and the #candidate_data
-             - In relevance to the #evaluation_type being assessed, look in the #answers for technical accuracy, depth of knowledge, problem-solving approaches, and communication skills.
-             - If the #answer is correct and the time of #answer is lower than the time given in the question, take it into account for positive assessment.
-             - Cross-reference #candidate answers with #job requirements to ensure role-specific #evaluation.
-             - Take evaluation type #coefficients in consideration
-             - Ensure fairness by matching evaluation difficulty and accuracy to candidate's stated experience level
-             - Use the exact 'description' value from each EvaluationTypes listFor the 'evaluationType' field in the output.
-             - Ensure each evaluation in the output array corresponds exactly to one EvaluationType from the input list.
-             - Do not skip any evaluation types or add additional ones not provided in the Evaluation Types.
-             - Reference specific technologies, skills, or experiences mentioned in the candidate profile when relevant.
-             - Use simple language: A2-B1-B2
-             - Return ONLY a valid JSON array with exactly this structure, no additional text or formatting:"{JSON_SCHEMA}", here you have a mock example:"{JSON_MOCK}".
-                                       
-            I provide bellow the needed information:
-             - #Candidate Profile: "{CANDIDATE_DATA}",
-             - #Job Offer requirements: "{OFFER_DATA}",
-             - #Evaluation Types criteria: "{EVALUATION_TYPES_DATA}".
-             - #Questions And Answers with the estimated answer time and the real answer time: "{QuestionAnswer_DATA}"
-            """;
-
-
-    @BeforeEach
-    void setUp() {
-        interviewEvaluationUtilServ = new InterviewEvaluationUtilServImpl(
-                evaluationServ,
-                interviewServ,
-                evaluationMapper,
-                promptServ
-        );
-    }
-
-    @Test
-    void should_return_saved_evaluations() throws Exception {
-        UUID interviewId = UUID.randomUUID();
-
-        QuestionsAndAnswersForEvaluationDTO questionsAndAnswersForEvaluation1 =
-                new QuestionsAndAnswersForEvaluationDTO(
-                        "what is the useState hook",
-                        "use state is",
-                        2,
-                        3);
-        QuestionsAndAnswersForEvaluationDTO questionsAndAnswersForEvaluation2 =
-                new QuestionsAndAnswersForEvaluationDTO(
-                        "what is the useEffect hook",
-                        "useEffect is",
-                        2,
-                        3);
-
-        PromptDTO prompt = new PromptDTO(UUID.randomUUID(),"test",INTERVIEW_EVALUATION_PROMPT,JSON_SCHEMA);
-
-        InterviewEvaluationsRequestDTO interviewEvaluationsRequest =
-                new InterviewEvaluationsRequestDTO("test",List.of(questionsAndAnswersForEvaluation1, questionsAndAnswersForEvaluation2));
-
-        Evaluation evaluation = new Evaluation();
-        evaluation.setFeedback("nothing");
-        evaluation.setScore(60.0);
-
-        EvaluationType evaluationType1 = new EvaluationType();
-        evaluationType1.setId(UUID.randomUUID());
-        evaluationType1.setDescription("React Skills");
-        evaluationType1.setCoefficient(3.0);
-
-        EvaluationType evaluationType2 = new EvaluationType();
-        evaluationType2.setId(UUID.randomUUID());
-        evaluationType2.setDescription("Java Skills");
-        evaluationType2.setCoefficient(2.0);
-
-        List<EvaluationType> evaluationTypes = List.of(evaluationType1, evaluationType2);
-
-        CandidateDTO candidateDTO = new CandidateDTO(
-                UUID.randomUUID(),
-                "John",
-                null,
-                10,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null,
-                null
-        );
-
-        OfferDTO offerDTO = new OfferDTO(
-                UUID.randomUUID(),
-                "Backend Developer",
-                "just testing",
-                null
-        );
-
-        PlaceholdersForInterviewEvaluationPromptDTO placeholders = new PlaceholdersForInterviewEvaluationPromptDTO(candidateDTO, offerDTO, evaluationTypes);
-
-
-        List<EvaluationsAIResponseDTO> aiEvaluationResponse = List.of(
-                new EvaluationsAIResponseDTO(4.0, "Good understanding of hooks", "react skills"),
-                new EvaluationsAIResponseDTO(3.5, "Basic understanding of lifecycle", "react skills")
-        );
-
-        List<Evaluation> existingEvaluations = List.of(evaluation);
-        List<Evaluation> savedEvaluations = List.of(evaluation);
-
-        when(interviewServ.getInterviewEvaluationPlaceholders(interviewId)).thenReturn(placeholders);
-        when(promptServ.findByPromptCode(interviewEvaluationsRequest.promptCode()))
-                .thenReturn(prompt);
-        when(evaluationServ.prepareEvaluationsDTOFromAiResponse(eq(interviewEvaluationsRequest), eq(placeholders),eq(prompt))).thenReturn(aiEvaluationResponse);
-        when(evaluationServ.getAllEvaluationsByInterviewID(interviewId)).thenReturn(existingEvaluations);
-        when(evaluationServ.saveAIEvaluationResponse(aiEvaluationResponse, existingEvaluations)).thenReturn(savedEvaluations);
-
-        List<Evaluation> result = interviewEvaluationUtilServ.prepareInterviewEvaluation(interviewId, interviewEvaluationsRequest);
-
-
-        assertEquals(savedEvaluations, result);
-        verify(interviewServ).getInterviewEvaluationPlaceholders(interviewId);
-        verify(evaluationServ).prepareEvaluationsDTOFromAiResponse(interviewEvaluationsRequest, placeholders, prompt);
-        verify(evaluationServ).getAllEvaluationsByInterviewID(interviewId);
-        verify(evaluationServ).saveAIEvaluationResponse(aiEvaluationResponse, existingEvaluations);
-    }
-
-
-    @Test
-    void should_return_interview_evaluations() throws Exception {
-
-        UUID interviewId = UUID.randomUUID();
-
-        EvaluationTypeDTO evaluationTypeDTO = new EvaluationTypeDTO(
-                UUID.randomUUID(),
-                "just testing",
-                3.0
-        );
-
-        FullEvaluationDTO fullEvaluationDTO = new FullEvaluationDTO(
-                UUID.randomUUID(),
-                60.0,
-                "just testing",
-                interviewId,
-                evaluationTypeDTO
-        );
-
-        LocalDateTime dateTime = LocalDateTime.now();
-        InterviewEvaluationDTO interviewEvaluation = new InterviewEvaluationDTO(
-                interviewId,
-                "test test",
-                "backend dev",
-                dateTime,
-                50,
-                List.of(fullEvaluationDTO)
-        );
-        Interview interview = new Interview();
-        interview.setId(interviewId);
-        interview.setScheduledAt(dateTime);
-
-        EvaluationType evaluationType = new EvaluationType();
-        evaluationType.setId(UUID.randomUUID());
-        evaluationType.setDescription("nothing");
-        evaluationType.setCoefficient(3.0);
-
-        Evaluation evaluation = new Evaluation(
-                UUID.randomUUID(),
-                60.0,
-                "just testing",
-                interview,
-                evaluationType
-        );
-
-        when(evaluationServ.getAllEvaluationsByInterviewID(interviewId)).thenReturn(List.of(evaluation));
-
-        when(evaluationMapper.mapEvaluationToInterviewEvaluation(List.of(evaluation)))
-                .thenReturn(interviewEvaluation);
-
-        InterviewEvaluationDTO result = interviewEvaluationUtilServ.getInterviewEvaluations(interviewId);
-
-        assertThat(result.scheduledAt()).isEqualTo(interview.getScheduledAt());
-
-        verify(evaluationServ).getAllEvaluationsByInterviewID(interviewId);
-        verify(evaluationMapper).mapEvaluationToInterviewEvaluation(List.of(evaluation));
-    }
-
-}
