+++ src/test/java/ma/nttdata/externals/module/interview/service/impl/InterviewSrvImplTest.java
@@ -1,5 +1,6 @@
 package ma.nttdata.externals.module.interview.service.impl;
 
+import ma.nttdata.externals.commons.constants.InterviewEvaluationPromptConstants;
 import ma.nttdata.externals.commons.exception.ResourceNotFoundException;
 import ma.nttdata.externals.module.candidate.dto.CandidateDTO;
 import ma.nttdata.externals.module.candidate.entity.Candidate;
@@ -9,21 +10,26 @@ import ma.nttdata.externals.module.interview.dto.*;
 import ma.nttdata.externals.module.interview.entity.*;
 import ma.nttdata.externals.module.interview.mapper.*;
 import ma.nttdata.externals.module.interview.repository.*;
+import ma.nttdata.externals.module.interview.service.EvaluationServ;
+import ma.nttdata.externals.module.interview.service.EvaluationTypeServ;
+import ma.nttdata.externals.module.interview.service.QuestionServ;
 import ma.nttdata.externals.module.offer.dto.OfferDTO;
 import ma.nttdata.externals.module.offer.entity.Offer;
 import ma.nttdata.externals.module.offer.mapper.OfferMapper;
 import ma.nttdata.externals.module.offer.repository.OfferRepository;
+import ma.nttdata.externals.module.prompt.dto.PromptDTO;
+import ma.nttdata.externals.module.prompt.service.PromptService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.springframework.data.crossstore.ChangeSetPersister;
 
 import java.time.LocalDateTime;
 import java.util.*;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.api.Assertions.in;
+import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
@@ -48,6 +54,10 @@ class InterviewSrvImplTest {
     private InterviewServImpl interviewServ;
 
     private static final String TEST_BASE_LINK = "http://localhost:8080/interview/";
+    @Mock private EvaluationServ evaluationServ;
+    @Mock private  PromptService promptServ;
+    @Mock private  EvaluationTypeServ evaluationTypeServ;
+    @Mock private QuestionServ questionServ;
 
     @BeforeEach
     void setUp() {
@@ -64,7 +74,11 @@ class InterviewSrvImplTest {
                 evaluationTypeMapper,
                 candidateMapper,
                 TEST_BASE_LINK,
-                offerMapper
+                offerMapper,
+                evaluationServ,
+                promptServ,
+                evaluationTypeServ,
+                questionServ
         );
     }
 
@@ -407,5 +421,156 @@ class InterviewSrvImplTest {
         assertTrue(interviewListDTOS.contains(interviewList));
         verify(interviewRepository).findAll();
         verify(interviewMapper).fromInterviewToInterviewListDTO(interview);
+    }
+
+    @Test
+    void should_return_saved_evaluations() throws Exception {
+        UUID interviewId = UUID.randomUUID();
+
+        QuestionsAndAnswersForEvaluationDTO questionsAndAnswersForEvaluation1 =
+                new QuestionsAndAnswersForEvaluationDTO(
+                        "what is the useState hook",
+                        "use state is",
+                        2,
+                        3);
+        QuestionsAndAnswersForEvaluationDTO questionsAndAnswersForEvaluation2 =
+                new QuestionsAndAnswersForEvaluationDTO(
+                        "what is the useEffect hook",
+                        "useEffect is",
+                        2,
+                        3);
+
+        PromptDTO prompt = new PromptDTO(UUID.randomUUID(),InterviewEvaluationPromptConstants.INTERVIEW_EVALUATION_PROMPT_CODE,InterviewEvaluationPromptConstants.INTERVIEW_EVALUATION_PROMPT,InterviewEvaluationPromptConstants.JSON_SCHEMA);
+
+        InterviewEvaluationsRequestDTO interviewEvaluationsRequest =
+                new InterviewEvaluationsRequestDTO(List.of(questionsAndAnswersForEvaluation1, questionsAndAnswersForEvaluation2));
+
+        Evaluation evaluation = new Evaluation();
+        evaluation.setFeedback("nothing");
+        evaluation.setScore(60.0);
+
+        EvaluationType evaluationType1 = new EvaluationType();
+        evaluationType1.setId(UUID.randomUUID());
+        evaluationType1.setDescription("React Skills");
+        evaluationType1.setCoefficient(3.0);
+
+        EvaluationType evaluationType2 = new EvaluationType();
+        evaluationType2.setId(UUID.randomUUID());
+        evaluationType2.setDescription("Java Skills");
+        evaluationType2.setCoefficient(2.0);
+
+        List<EvaluationType> evaluationTypes = List.of(evaluationType1, evaluationType2);
+
+        CandidateDTO candidateDTO = new CandidateDTO(
+                UUID.randomUUID(),
+                "John",
+                null,
+                10,
+                null,
+                null,
+                null,
+                null,
+                null,
+                null,
+                null,
+                null,
+                null,
+                null,
+                null
+        );
+
+        OfferDTO offerDTO = new OfferDTO(
+                UUID.randomUUID(),
+                "Backend Developer",
+                "just testing",
+                null
+        );
+
+        PlaceholdersForInterviewEvaluationPromptDTO placeholders = new PlaceholdersForInterviewEvaluationPromptDTO(candidateDTO, offerDTO, evaluationTypes);
+
+
+        List<EvaluationsAIResponseDTO> aiEvaluationResponse = List.of(
+                new EvaluationsAIResponseDTO(4.0, "Good understanding of hooks", "react skills"),
+                new EvaluationsAIResponseDTO(3.5, "Basic understanding of lifecycle", "react skills")
+        );
+
+        List<Evaluation> existingEvaluations = List.of(evaluation);
+        List<Evaluation> savedEvaluations = List.of(evaluation);
+        InterviewServImpl interviewServSpy = Mockito.spy(interviewServ);
+
+        doReturn(placeholders).when(interviewServSpy).getInterviewEvaluationPlaceholders(interviewId);
+        when(promptServ.findByPromptCode(InterviewEvaluationPromptConstants.INTERVIEW_EVALUATION_PROMPT_CODE))
+                .thenReturn(prompt);
+        when(evaluationServ.prepareEvaluationsDTOFromAiResponse(eq(interviewEvaluationsRequest), eq(placeholders),eq(prompt))).thenReturn(aiEvaluationResponse);
+        when(evaluationServ.getAllEvaluationsByInterviewID(interviewId)).thenReturn(existingEvaluations);
+        when(evaluationServ.saveAIEvaluationResponse(aiEvaluationResponse, existingEvaluations)).thenReturn(savedEvaluations);
+
+        List<Evaluation> result = interviewServSpy.prepareInterviewEvaluation(interviewId, interviewEvaluationsRequest);
+
+
+        assertEquals(savedEvaluations, result);
+        verify(interviewServSpy).getInterviewEvaluationPlaceholders(interviewId);
+        verify(evaluationServ).prepareEvaluationsDTOFromAiResponse(interviewEvaluationsRequest, placeholders, prompt);
+        verify(evaluationServ).getAllEvaluationsByInterviewID(interviewId);
+        verify(evaluationServ).saveAIEvaluationResponse(aiEvaluationResponse, existingEvaluations);
+    }
+
+
+    @Test
+    void should_return_interview_evaluations() throws Exception {
+
+        UUID interviewId = UUID.randomUUID();
+
+        EvaluationTypeDTO evaluationTypeDTO = new EvaluationTypeDTO(
+                UUID.randomUUID(),
+                "just testing",
+                3.0
+        );
+
+        FullEvaluationDTO fullEvaluationDTO = new FullEvaluationDTO(
+                UUID.randomUUID(),
+                60.0,
+                "just testing",
+                interviewId,
+                evaluationTypeDTO
+        );
+
+        LocalDateTime dateTime = LocalDateTime.now();
+        InterviewEvaluationDTO interviewEvaluation = new InterviewEvaluationDTO(
+                interviewId,
+                "test test",
+                "backend dev",
+                dateTime,
+                50,
+                List.of(fullEvaluationDTO)
+        );
+        Interview interview = new Interview();
+        interview.setId(interviewId);
+        interview.setScheduledAt(dateTime);
+
+        EvaluationType evaluationType = new EvaluationType();
+        evaluationType.setId(UUID.randomUUID());
+        evaluationType.setDescription("nothing");
+        evaluationType.setCoefficient(3.0);
+
+        Evaluation evaluation = new Evaluation(
+                UUID.randomUUID(),
+                60.0,
+                "just testing",
+                interview,
+                evaluationType
+        );
+
+        when(evaluationServ.getAllEvaluationsByInterviewID(interviewId)).thenReturn(List.of(evaluation));
+
+        when(evaluationMapper.mapEvaluationToInterviewEvaluation(List.of(evaluation)))
+                .thenReturn(interviewEvaluation);
+
+        InterviewEvaluationDTO result = interviewServ.getInterviewEvaluations(interviewId);
+
+        assertThat(result.scheduledAt()).isEqualTo(interview.getScheduledAt());
+
+        verify(evaluationServ).getAllEvaluationsByInterviewID(interviewId);
+        verify(evaluationMapper).mapEvaluationToInterviewEvaluation(List.of(evaluation));
     }
 }
